refactor: use tuple instead of field and move fail logic to testing-utils
