+++ crates/task-manager/src/response_calculator.rs
@@ -1,6 +1,4 @@
-use rand::Rng;
-use std::{future::Future, marker::PhantomData};
-use tracing::info;
+use std::future::Future;
 
 use crate::TaskManagerError;
 
@@ -25,40 +23,9 @@ pub trait ResponseCalculator<Input, Output> {
 
 /// Implementation of the [`ResponseCalculator`] trait that uses a function to compute the response.
 #[derive(Debug)]
-pub struct FunctionResponseCalculator<Input, Output, F>
-where
-    F: AsyncFn(u32, Input) -> Result<Output, TaskManagerError>,
-{
-    /// The function used to compute the response.
-    pub compute_fn: F,
-    _input: PhantomData<Input>,
-    _output: PhantomData<Output>,
-}
+pub struct FunctionResponseCalculator<F>(F);
 
-impl<Input, Output, F> FunctionResponseCalculator<Input, Output, F>
-where
-    F: AsyncFn(u32, Input) -> Result<Output, TaskManagerError>,
-{
-    /// Create a new `FunctionResponseCalculator`.
-    ///
-    /// # Arguments
-    ///
-    /// * `compute_fn` - The function used to compute the response.
-    ///
-    /// # Returns
-    ///
-    /// * `FunctionResponseCalculator<Input, Output, F>` - A new `FunctionResponseCalculator`.
-    pub fn new(compute_fn: F) -> Self {
-        Self {
-            compute_fn,
-            _input: PhantomData,
-            _output: PhantomData,
-        }
-    }
-}
-
-impl<Input, Output, F> ResponseCalculator<Input, Output>
-    for FunctionResponseCalculator<Input, Output, F>
+impl<F, Input, Output> ResponseCalculator<Input, Output> for FunctionResponseCalculator<F>
 where
     F: AsyncFn(u32, Input) -> Result<Output, TaskManagerError>,
 {
@@ -67,54 +34,43 @@ where
         task_index: u32,
         input: Input,
     ) -> Result<Output, TaskManagerError> {
-        (self.compute_fn)(task_index, input).await
+        (self.0)(task_index, input).await
     }
 }
 
-/// Helper to wrap both correct and incorrect logic in a single closure.
-/// USE THIS FOR TESTING PURPOSES ONLY
-///
-/// # Arguments
-///
-/// * `response_calculator` - The response calculator to use.
-/// * `incorrect_logic` - The incorrect logic to respond to the task.
-/// * `failure_rate_percentage` - The failure rate percentage.
-///
-/// # Returns
-///
-/// * `impl ResponseCalculator<Input, Output>` - The wrapped logic.
-///
-/// # Panics
-///
-/// Panics if `failure_rate_percentage` is greater than 100.
-#[cfg(feature = "operator-testing")]
-pub fn failing_response_calculator<Input, Output>(
-    response_calculator: impl ResponseCalculator<Input, Output>,
-    incorrect_logic: impl AsyncFn(u32, Input) -> Result<Output, TaskManagerError>,
-    failure_rate_percentage: u32,
-) -> impl ResponseCalculator<Input, Output>
-where
-    Input: Clone,
-{
-    assert!(
-        failure_rate_percentage <= 100,
-        "Failure rate percentage must be less than or equal to 100"
-    );
+impl<F> FunctionResponseCalculator<F> {
+    /// Create a new [`FunctionResponseCalculator`] from a sync function.
+    /// This function will be converted to an async function.
+    ///
+    /// # Arguments
+    ///
+    /// * `compute_fn` - The sync function to compute the response.
+    ///
+    /// # Returns
+    ///
+    /// * [`FunctionResponseCalculator`] - The new [`FunctionResponseCalculator`].
+    pub fn new<CF, Input, Output>(
+        compute_fn: CF,
+    ) -> FunctionResponseCalculator<impl AsyncFn(u32, Input) -> Result<Output, TaskManagerError>>
+    where
+        CF: Fn(u32, Input) -> Result<Output, TaskManagerError>,
+    {
+        FunctionResponseCalculator(async move |a, b| compute_fn(a, b))
+    }
 
-    FunctionResponseCalculator::new(async move |task_index, input: Input| {
-        let result = response_calculator
-            .compute_response(task_index, input.clone())
-            .await;
-
-        let mut rng = rand::thread_rng();
-        let should_fail = rng.gen_bool(failure_rate_percentage as f64 / 100.0);
-
-        if should_fail {
-            info!("Operator compute the task with a wrong response");
-            incorrect_logic(task_index, input).await
-        } else {
-            info!("Operator compute the task successfully");
-            result
-        }
-    })
+    /// Create a new [`FunctionResponseCalculator`] from an async function.
+    ///
+    /// # Arguments
+    ///
+    /// * `compute_fn` - The async function to compute the response.
+    ///
+    /// # Returns
+    ///
+    /// * [`FunctionResponseCalculator`] - The new [`FunctionResponseCalculator`].
+    pub fn new_async<CF, Input, Output>(compute_fn: CF) -> FunctionResponseCalculator<CF>
+    where
+        CF: AsyncFn(u32, Input) -> Result<Output, TaskManagerError>,
+    {
+        FunctionResponseCalculator(compute_fn)
+    }
 }
