+++ tests/testing-utils/Cargo.toml
@@ -18,7 +18,10 @@ eigen-client-elcontracts.workspace = true
 eigen-common.workspace = true
 eigen-utils.workspace = true
 eigen-crypto-bls.workspace = true
+eigen-task-manager.workspace = true
+rand.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+tracing.workspace = true
 testcontainers.workspace = true
 url.workspace = true
